Issue #61 / [86133850] committing code for adding active indicator and removing delete from batch correspondence document. Moved the active indicator from BatchCorrespondence to BatchCorrespondenceBase.
