--- src/main/java/org/kuali/kra/irb/correspondence/BatchCorrespondence.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2005-2014 The Kuali Foundation
- *
- * Licensed under the Educational Community License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.osedu.org/licenses/ECL-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.kuali.kra.irb.correspondence;
-
-import org.kuali.kra.protocol.correspondence.BatchCorrespondenceBase;
-
-public class BatchCorrespondence extends BatchCorrespondenceBase {
-
-    /**
-     * Comment for <code>serialVersionUID</code>
-     */
-    private static final long serialVersionUID = 3467000376519170931L;
-    
-//  ### Vivantech Fix : #61 / [#86133850] adding active indicator field and disabling the delete.
-    private boolean active;
-
-	public boolean isActive() {
-		return active;
-	}
-
-	public void setActive(boolean active) {
-		this.active = active;
-	}
-}
